Make sure rom gets serialized as a file path
Previously state files included the whole running rom as a binary blob.
Now they include a path to the rom instead.

Eventually this should also include a hash of the original file, and there
should be a way to specify an alternate path for the file. Also the path
should be relative, not absolute.